+++ src/main/kotlin/db/tables/Messages.kt
@@ -395,7 +395,7 @@ object Messages : IntIdTable() {
      *
      * @see [Messages.exists]
      */
-    fun readChatFromMessage(messageId: Int): Int = transaction {
+    fun readChatIdFromMessageId(messageId: Int): Int = transaction {
         select { Messages.id eq messageId }.first()[chatId]
     }
 
@@ -464,7 +464,7 @@ object Messages : IntIdTable() {
      * [messagesNotifier].
      */
     fun delete(id: Int) {
-        val chatId = readChatFromMessage(id)
+        val chatId = readChatIdFromMessageId(id)
         deleteChatMessages(id)
         messagesNotifier.publish(DeletedMessage(chatId, id), readUserIdList(chatId))
     }
@@ -554,14 +554,18 @@ object Messages : IntIdTable() {
      *
      * @return
      * - `false` if the [messageId] doesn't exist.
+     * - `true` if the [messageId] is from a public group chat (regardless of whether the [userId] is a participant).
+     * - `false` if the [userId] is `null`, and the [messageId] isn't from a public group chat.
      * - `false` if the [messageId] is from a chat the user isn't in.
      * - `true` if the [messageId] is from a group chat the [userId] is in.
      * - `true` if the [messageId] is from a private chat the [userId] hasn't deleted.
      * - `false` if the [messageId] was sent before the [userId] deleted the private chat.
      */
-    fun isVisible(userId: Int, messageId: Int): Boolean {
+    fun isVisible(userId: Int?, messageId: Int): Boolean {
         if (!exists(messageId)) return false
-        val chatId = readChatFromMessage(messageId)
+        val chatId = readChatIdFromMessageId(messageId)
+        if (GroupChats.isExistentPublicChat(chatId)) return true
+        if (userId == null) return false
         if (!isUserInChat(userId, chatId)) return false
         if (chatId in GroupChatUsers.readChatIdList(userId)) return true
         val deletion = PrivateChatDeletions.readLastDeletion(chatId, userId) ?: return true