+++ src/test/kotlin/restApi/PicMessageTest.kt
@@ -13,7 +13,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-private fun getPicMessage(accessToken: String, messageId: Int, type: PicType): TestApplicationResponse =
+private fun getPicMessage(accessToken: String? = null, messageId: Int, type: PicType): TestApplicationResponse =
     getFileMessage(accessToken, path = "pic-message", messageId, type)
 
 private fun postPicMessage(
@@ -60,9 +60,9 @@ class PicMessageTest {
         }
 
         @Test
-        fun `An HTTP status code of 400 must be returned when retrieving a nonexistent message`() {
+        fun `An HTTP status code of 401 must be returned when retrieving a nonexistent message`() {
             val token = createVerifiedUsers(1)[0].accessToken
-            assertEquals(HttpStatusCode.BadRequest, getPicMessage(token, messageId = 1, PicType.ORIGINAL).status())
+            assertEquals(HttpStatusCode.Unauthorized, getPicMessage(token, messageId = 1, PicType.ORIGINAL).status())
         }
 
         private fun createMessage(): Pair<String, Int> {